requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
bidict==0.23.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
fastapi==0.120.4
h11==0.16.0
idna==3.11
pydantic==2.12.3
pydantic_core==2.41.4
python-dotenv
python-engineio==3.14.2
python-socketio==4.6.1
requests==2.32.5
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sqlalchemy
starlette==0.49.2
typing_extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.5.0
uvicorn[standard]==0.38.0
websocket-client==1.9.0
wsproto==1.3.1
alembic
pytest
httpx